--- src/main/java/com/teamrocket/Template/OrderServiceMain.java
@@ -1,13 +1,0 @@
-package com.teamrocket.Template;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OrderServiceMain {
-
-	public static void main(String[] args) {
-		SpringApplication.run(OrderServiceMain.class, args);
-	}
-
-}
