--- src/main/java/com/teamrocket/Template/dto/NewOrderItem.java
@@ -1,15 +1,0 @@
-package com.teamrocket.Template.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-public class NewOrderItem {
-    private int menuItemId;
-    private int quantity;
-}
