--- src/main/java/com/teamrocket/Template/dto/OrderStatusDTO.java
@@ -1,14 +1,0 @@
-package com.teamrocket.Template.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-public class OrderStatusDTO {
-    private String status;
-}
