--- src/main/java/com/teamrocket/Template/enums/OrderStatus.java
@@ -1,5 +1,0 @@
-package com.teamrocket.Template.enums;
-
-public enum OrderStatus {
-    PENDING, IN_PROGRESS, READY, PICKED_UP, COMPLETED, CANCELED;
-}
