--- src/main/java/com/teamrocket/Template/repository/OrderRepository.java
@@ -1,10 +1,0 @@
-package com.teamrocket.Template.repository;
-
-import com.teamrocket.Template.entity.Order;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface OrderRepository extends JpaRepository<Order, Integer> {
-
-}
