--- src/main/java/com/teamrocket/Template/service/TemplateService.java
@@ -1,14 +1,0 @@
-package com.teamrocket.Template.service;
-
-import com.teamrocket.Template.dto.TemplateDTO;
-import org.springframework.stereotype.Service;
-import com.teamrocket.Template.control.TempController;
-
-@Service
-public class TemplateService {
-
-    // TODO: DELETEME
-    public TemplateDTO hello(String who) {
-        return new TemplateDTO(99, "Hello, " + who + "!");
-    }
-}
