+++ src/main/java/com/teamrocket/orderservice/OrderServiceMain.java
@@ -1,0 +1,13 @@
+package com.teamrocket.orderservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OrderServiceMain {
+
+	public static void main(String[] args) {
+		SpringApplication.run(OrderServiceMain.class, args);
+	}
+
+}
