+++ src/main/java/com/teamrocket/orderservice/dto/NewOrderItem.java
@@ -1,0 +1,15 @@
+package com.teamrocket.orderservice.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+public class NewOrderItem {
+    private int menuItemId;
+    private int quantity;
+}
