+++ src/main/java/com/teamrocket/orderservice/dto/OrderStatusDTO.java
@@ -1,0 +1,14 @@
+package com.teamrocket.orderservice.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+public class OrderStatusDTO {
+    private String status;
+}
