+++ src/main/java/com/teamrocket/orderservice/enums/OrderStatus.java
@@ -1,0 +1,5 @@
+package com.teamrocket.orderservice.enums;
+
+public enum OrderStatus {
+    PENDING, IN_PROGRESS, READY, PICKED_UP, COMPLETED, CANCELED;
+}
