+++ src/main/java/com/teamrocket/orderservice/repository/OrderRepository.java
@@ -1,0 +1,10 @@
+package com.teamrocket.orderservice.repository;
+
+import com.teamrocket.orderservice.entity.Order;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface OrderRepository extends JpaRepository<Order, Integer> {
+
+}
