--- src/test/java/com/teamrocket/Template/OrderServiceMainTests.java
@@ -1,13 +1,0 @@
-package com.teamrocket.Template;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class OrderServiceMainTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
