--- src/test/java/com/teamrocket/Template/acceptance/TempDef.java
@@ -1,11 +1,0 @@
-package com.teamrocket.Template.acceptance;
-
-import io.cucumber.java.en.Given;
-import io.cucumber.java.en.Then;
-import io.cucumber.java.en.When;
-
-import static org.junit.Assert.assertTrue;
-
-public class TempDef {
-
-}
