--- src/test/java/com/teamrocket/Template/acceptance/config/CucumberAcceptanceTest.java
@@ -1,21 +1,0 @@
-package com.teamrocket.Template.acceptance.config;
-
-import io.cucumber.junit.platform.engine.Constants;
-import org.junit.platform.suite.api.ConfigurationParameter;
-import org.junit.platform.suite.api.IncludeEngines;
-import org.junit.platform.suite.api.SelectClasspathResource;
-import org.junit.platform.suite.api.Suite;
-
-@Suite
-@IncludeEngines("cucumber")
-@SelectClasspathResource("features")
-@ConfigurationParameter(key = Constants.PLUGIN_PROPERTY_NAME, value = "pretty")
-@ConfigurationParameter(
-        key = Constants.GLUE_PROPERTY_NAME,
-        value = "com.teamrocket.Template.acceptance," +
-                "com.teamrocket.Template.acceptance.config"
-)
-@ConfigurationParameter(key = Constants.PLUGIN_PUBLISH_QUIET_PROPERTY_NAME, value = "true")
-public class CucumberAcceptanceTest {
-
-}
