--- src/test/java/com/teamrocket/Template/acceptance/config/CucumberSpringContextConfig.java
@@ -1,12 +1,0 @@
-package com.teamrocket.Template.acceptance.config;
-
-import io.cucumber.spring.CucumberContextConfiguration;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-@AutoConfigureMockMvc
-@CucumberContextConfiguration
-public class CucumberSpringContextConfig {
-
-}
