+++ src/test/java/com/teamrocket/orderservice/OrderServiceMainTests.java
@@ -1,0 +1,13 @@
+package com.teamrocket.orderservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class OrderServiceMainTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
