+++ src/test/java/com/teamrocket/orderservice/acceptance/TempDef.java
@@ -1,0 +1,7 @@
+package com.teamrocket.orderservice.acceptance;
+
+import static org.junit.Assert.assertTrue;
+
+public class TempDef {
+
+}
