+++ src/test/java/com/teamrocket/orderservice/acceptance/config/CucumberSpringContextConfig.java
@@ -1,0 +1,12 @@
+package com.teamrocket.orderservice.acceptance.config;
+
+import io.cucumber.spring.CucumberContextConfiguration;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+@CucumberContextConfiguration
+public class CucumberSpringContextConfig {
+
+}
